feat: complete Result API and add workflow classes
Add missing Rust-style Result methods and workflow inheritance classes.
Bump version to 0.1.2.
