+++ src/neopipe/__about__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2024-present MrDataPsycho <mr.data.psycho@gmail.com>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.1.1"
+__version__ = "0.1.2"
