+++ src/neopipe/__init__.py
@@ -7,14 +7,25 @@ logging.basicConfig(
     datefmt="%Y-%m-%d %H:%M:%S",
 )
 
-# # Import necessary modules
-from .result import Err, Ok, Result
-
-# from .task import AsyncTask, SyncTask, AbstractAsyncTask, AbstractSyncTask
-# from .pipeline import Pipeline
-
+# Import core components
+from neopipe.result import Err, Ok, Result, ExecutionResult, Trace, Traces
+from neopipe.task import (
+    BaseSyncTask, 
+    BaseAsyncTask,
+    FunctionSyncTask,
+    FunctionAsyncTask,
+    ClassSyncTask,
+    ClassAsyncTask
+)
+from neopipe.pipeline import SyncPipeline, AsyncPipeline
+from neopipe.workflow import SyncWorkflow, AsyncWorkflow
 
 # Specify what is available for import from this package
-__all__ = ["Result", "Ok", "Err"]
+__all__ = [
+    "Result", "Ok", "Err", "ExecutionResult", "Trace", "Traces",
+    "BaseSyncTask", "BaseAsyncTask", "FunctionSyncTask", "FunctionAsyncTask", 
+    "ClassSyncTask", "ClassAsyncTask",
+    "SyncPipeline", "AsyncPipeline", "SyncWorkflow", "AsyncWorkflow"
+]
 
 from .__about__ import __version__
